Consolidate common functions into utils.py
