+++ nodewatcher/nodewatcher.cfg
@@ -3,3 +3,4 @@ region = us-east-1
 scheduler = test
 proxy = NONE
 scaledown_idletime = 10
+stack_name = cfncluster-test
